\name{Affyhgu133aExpr-package}
\alias{Affyhgu133aExpr-package}
\docType{package}
\title{
Affymetrix Human hgu133a Array (GPL96) Expression Data Package
}
\description{
Contains gene expression profiles from Affymetrix Human hgu133a Array (GPL96). This package is primarily designed for GSCA (Gene Set Context Analysis). All gene expression data are downloaded from NCBI GEO. Gene expression data were preprocessed and normalized consistently using fRMA. Gene expression values are further standardized using gene expression barcode approach. The biological context of each sample was recorded and manually verified based on the sample description in GEO. Gene expression profiles are stored as hdf5 format.
}
\details{
\tabular{ll}{
Package: \tab Affyhgu133aExpr\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2014-4-9\cr
License: \tab GPL 2.0\cr
}
}
\author{
Author: Zhicheng Ji, Hongkai Ji
Maintainer: Zhicheng Ji <zji4@jhu.edu>
}
\references{
McCall M.N., Bolstad B.M., and Irizarry R.A. (2010) Frozen robust multiarray analysis (fRMA). Biostatistics 11, 242-253.

McCall, M. N., Uppal, K., Jaffee, H. A., Zilliox, M. J., & Irizarry, R. A. (2011). The Gene Expression Barcode: leveraging public data repositories to begin cataloging the human and murine transcriptomes. Nucleic acids research, 39(suppl 1), D1011-D1015.

Barrett T., et al. (2007) NCBI GEO: mining tens of millions of expression profiles - database and tools update. Nucl. Acids Res. 35, D760-D765.
}
\keyword{ package, database, Affyhgu133aExpr }

